# docs/swagger.yaml
definitions:
  dto.LoginDto:
    properties:
      password:
        description: 密码
        type: string
      userName:
        description: 用户账号
        type: string
    type: object
  dto.SysUserAddDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      nickName:
        description: 用户昵称
        type: string
      password:
        description: 密码
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
      userName:
        description: 用户账号
        type: string
    type: object
  dto.SysUserUpdateDto:
    properties:
      email:
        description: 用户邮箱
        type: string
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
    type: object
  dto.SysUserVo:
    properties:
      avatar:
        description: 头像地址
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: 用户邮箱
        type: string
      id:
        description: 编号
        type: integer
      nickName:
        description: 用户昵称
        type: string
      phone:
        description: 手机号码
        type: string
      remark:
        description: 备注
        type: string
      sex:
        description: 用户性别（0未知，1男，2女）
        type: integer
      status:
        description: 帐号状态（0正常 1停用）
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      userName:
        description: 用户账号
        type: string
      userType:
        description: 用户类型（0普通账号，1超级管理员）
        type: integer
    type: object
  request.IdInfoDto:
    properties:
      id:
        description: id编号
        type: integer
    type: object
  response.JsonResult:
    properties:
      code:
        description: 状态码（架构层面：中间件/异常处理等）
        type: integer
      data:
        description: 数据
      message:
        description: 消息
        type: string
      success:
        description: 是否成功（业务层面）
        type: boolean
    type: object
  response.PageResult:
    properties:
      list:
        description: 数据
      totalCount:
        description: 总条数
        type: integer
    type: object
info:
  contact: {}
  description: go-fast-admin 接口文档
  title: 系统模块
  version: 1.0.0
paths:
  /system/auth/login:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 授权
  /system/user/add:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserAddDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 添加用户
      tags:
      - 用户
  /system/user/delete:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdInfoDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - 用户
  /system/user/detail:
    get:
      parameters:
      - description: id编号
        in: query
        name: id
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  $ref: '#/definitions/dto.SysUserVo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取用户详情
      tags:
      - 用户
  /system/user/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.SysUserVo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户列表
      tags:
      - 用户
  /system/user/query:
    get:
      parameters:
      - description: 页码
        in: query
        name: pageIndex
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户昵称
        in: query
        name: userName
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResult'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResult'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.SysUserVo'
                        type: array
                    type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用户查询
      tags:
      - 用户
  /system/user/update:
    post:
      parameters:
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysUserUpdateDto'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JsonResult'
      security:
      - ApiKeyAuth: []
      summary: 更新用户
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"